+++ src/main/java/dev/ftb/packcompanion/PackCompanion.java
@@ -8,8 +8,10 @@ import dev.ftb.packcompanion.core.Feature;
 import dev.ftb.packcompanion.features.forcedgamemodes.ForcedGameModesFeature;
 import dev.ftb.packcompanion.features.forcedgamerule.ForcedGameRulesFeature;
 import dev.ftb.packcompanion.features.onboarding.shadernotice.ShaderNotice;
+import dev.ftb.packcompanion.features.reciperemover.RecipeRemover;
 import dev.ftb.packcompanion.features.schematic.SchematicPasteFeature;
 import dev.ftb.packcompanion.features.spawners.SpawnerFeature;
+import dev.ftb.packcompanion.features.structureplacer.StructurePlacerFeature;
 import dev.ftb.packcompanion.features.structures.StructuresFeature;
 import dev.ftb.packcompanion.features.actionpad.ActionPadFeature;
 import dev.ftb.packcompanion.features.triggerblock.TriggerBlockFeature;
@@ -61,7 +63,9 @@ public class PackCompanion {
             ActionPadFeature::new,
             ForcedGameRulesFeature::new,
             ForcedGameModesFeature::new,
-            SchematicPasteFeature::new
+            SchematicPasteFeature::new,
+            StructurePlacerFeature::new,
+            RecipeRemover::new
     );
 
     private final List<Feature> createdFeatures = new ArrayList<>();